Fixing issue with if statement thanks to Tom Wood!
The problem you are running into seems to be handling the case where trafilatura.fetch_url returns an empty string (which is apparently documented API).  The empty string is not the same as None, and it looks like you can't build a readability Document() object from an empty string.  Hence the boom.